Add DispatchedNotification class
The DispatchedNotification class is a wrapper for Notification instances after they have been dispatched / sent to the notifications server.

Instances of DispatchedNotification have their own unique identifier whose format varies from platform to platform. For example, on Linux it's a simple uint32 converted to str. On other platforms the identifier might be identical to the randomly generated UUID of the wrapped Notification instance, but users must not rely on this.

The DispatchedNotification class furthermore exposes the status of the notification, namely whether it was cleared, clicked, dismissed, replied to, or what button was clicked.

The DesktopNotifier.send_notification() method now returns a DispatchedNotification instance. It also accepts DispatchedNotification instances to send a previously sent notification again. If a DispatchedNotification instance is given and that notification is still active, DesktopNotifier will attempt to replace that notification by the new one (on Linux this works seamlessly, on other platforms the previous notification is simply closed and a new notification is sent).

The DesktopNotifier.get_current_notifications() method (as well as their backend counterpart) now returns a dict of DispatchedNotification instances indexed by their identifier.