Add DispatchedNotification class
The `DispatchedNotification` class is a wrapper for `Notification` instances after they have been dispatched / sent to the notifications server.

Instances of `DispatchedNotification` have their own unique identifier whose format varies from platform to platform. For example, on Linux it's a simple uint32 converted to str. On other platforms the identifier might be identical to the randomly generated UUID of the wrapped Notification instance, but users must not rely on this. For example, if the same `Notification` instance is sent twice, the second `DispatchedNotification` will have a different identifier.

The `DispatchedNotification` class furthermore exposes the status of the notification, namely whether it was cleared, clicked, dismissed, replied to, or what button was clicked.

The `DesktopNotifier.send_notification()` method now returns a `DispatchedNotification` instance. It also accepts `DispatchedNotification` instances to send a previously sent notification again (just create a new `DispatchedNotification` instance with the same `identifier`, but a different `notification`). If a `DispatchedNotification` instance is given and that notification is still active, Desktop Notifier will attempt to replace that notification by the new one (on Linux this works seamlessly, on other platforms the previous notification is closed first and a new notification is sent).

Since some platforms (currently just OS X) support returning the "actual" list of active notifications, `DesktopNotifier.get_current_notifications()` doesn't return `DispatchedNotification` instances, but still just a list of the identifiers, matching those of `DispatchedNotification` instances. To access these there's a new `DesktopNotifier.get_cached_notifications()` method.